requirements.txt
streamlit
transformers
opencv-python
pillow